refactor: move `DataQualityEvaluator` to `evals/features/common.py`
Extract `DataQualityEvaluator` from `test_suggest_impact.py` into the
shared `evals/features/common.py` module, generalized from
`Evaluator[str, SuggestImpactModel]` to `Evaluator[str, AegisFeatureModel]`
so it can be reused across eval suites.

Related: https://redhat.atlassian.net/browse/AEGIS-490
